requirements.txt
# ============================================
# Python 项目依赖
# ============================================

# 环境变量管理
python-dotenv>=1.0.0

# 网络请求
requests>=2.31.0

# 数据处理
pandas>=2.0.0

# LLM API (DeepSeek / 智谱 AI)
openai>=1.0.0

# 阿里云服务
dashscope>=1.20.0
oss2>=2.18.0

# 可选：音频处理（如需合并音频）
# pydub>=0.25.0  # 需要安装 ffmpeg
